etc: 코드 포맷팅 및 필요없는 메서드 삭제
