+++ backend/src/test/java/botobo/core/admin/AdminAcceptanceTest.java
@@ -1,15 +1,10 @@
 package botobo.core.admin;
 
-import botobo.core.admin.dto.AdminAnswerRequest;
-import botobo.core.admin.dto.AdminAnswerResponse;
-import botobo.core.AcceptanceTest;
 import botobo.core.admin.dto.AdminCardRequest;
 import botobo.core.admin.dto.AdminCardResponse;
-import botobo.core.admin.dto.AdminCategoryRequest;
-import botobo.core.admin.dto.AdminCategoryResponse;
-import botobo.core.auth.AuthAcceptanceTest;
 import botobo.core.admin.dto.AdminWorkbookRequest;
 import botobo.core.admin.dto.AdminWorkbookResponse;
+import botobo.core.auth.AuthAcceptanceTest;
 import botobo.core.exception.ErrorResponse;
 import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
@@ -18,8 +13,6 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
-
-import java.util.List;
 
 import static botobo.core.TestUtils.extractId;
 import static botobo.core.TestUtils.longStringGenerator;
@@ -31,8 +24,7 @@ public class AdminAcceptanceTest extends AuthAcceptanceTest {
     private static final AdminWorkbookRequest ADMIN_CATEGORY_REQUEST =
             new AdminWorkbookRequest("관리자의 문제집");
 
-    public ExtractableResponse<Response> 카테고리_생성_요청(AdminCategoryRequest adminCategoryRequest) {
-    public static ExtractableResponse<Response> 문제집_생성_요청(AdminWorkbookRequest adminWorkbookRequest) {
+    public ExtractableResponse<Response> 문제집_생성_요청(AdminWorkbookRequest adminWorkbookRequest) {
         return RestAssured.given().log().all()
                 .contentType(MediaType.APPLICATION_JSON_VALUE)
                 .auth().oauth2(로그인되어_있음().getAccessToken())
@@ -40,18 +32,6 @@ public class AdminAcceptanceTest extends AuthAcceptanceTest {
                 .when().post("/admin/workbooks")
                 .then().log().all()
                 .extract();
-    }
-
-    public static void 여러개_문제집_생성_요청(List<AdminWorkbookRequest> requests) {
-        for (AdminWorkbookRequest request : requests) {
-            RestAssured.given().log().all()
-                    .contentType(MediaType.APPLICATION_JSON_VALUE)
-                    .auth().oauth2(로그인되어_있음().getAccessToken())
-                    .body(request)
-                    .when().post("/admin/workbooks")
-                    .then().log().all()
-                    .extract();
-        }
     }
 
     public ExtractableResponse<Response> 카드_생성_요청(AdminCardRequest adminCardRequest) {
@@ -64,17 +44,6 @@ public class AdminAcceptanceTest extends AuthAcceptanceTest {
                 .extract();
     }
 
-    public void 여러개_카드_생성_요청(List<AdminCardRequest> requests) {
-        for (AdminCardRequest request : requests) {
-            RestAssured.given().log().all()
-                    .contentType(MediaType.APPLICATION_JSON_VALUE)
-                    .auth().oauth2(로그인되어_있음().getAccessToken())
-                    .body(request)
-                    .when().post("/admin/cards")
-                    .then().log().all()
-                    .extract();
-        }
-    }
 
     @Test
     @DisplayName("관리자 문제집 생성 - 성공")
