+++ backend/src/test/java/botobo/core/documentation/AdminDocumentationTest.java
@@ -6,10 +6,8 @@ import botobo.core.admin.dto.AdminCardResponse;
 import botobo.core.admin.dto.AdminWorkbookRequest;
 import botobo.core.admin.dto.AdminWorkbookResponse;
 import botobo.core.admin.ui.AdminController;
-import botobo.core.quiz.exception.WorkbookNotFoundException;
 import botobo.core.auth.infrastructure.JwtTokenProvider;
-import botobo.core.quiz.exception.CardNotFoundException;
-import botobo.core.quiz.exception.CategoryNotFoundException;
+import botobo.core.quiz.exception.WorkbookNotFoundException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
